ai-job-applier-desktop/backend/requirements.txt
# ========================================
# AI求职助手桌面版 - Python依赖包
# ========================================

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
websockets==12.0

# 环境配置
python-dotenv==1.0.0

# AI接口
openai==1.10.0

# HTTP请求
requests>=2.32.0
aiohttp>=3.9.0

# 文档解析
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl>=3.1.0

# 数据处理
pandas>=2.0.0

# 浏览器自动化
playwright==1.41.0
playwright-stealth==1.0.0
selenium>=4.15.0
undetected-chromedriver>=3.5.4
webdriver-manager>=4.0.1
DrissionPage>=4.0.0

# 验证码识别（可选）
ddddocr>=1.4.0

# 图片处理
Pillow>=10.4.0
pytesseract>=0.3.10

# 定时任务
schedule>=1.2.0

# 性能监控
psutil>=5.9.0

# 数据库
sqlalchemy==2.0.25
aiofiles==23.2.1

# 缓存
# redis>=5.0.0  # 可选

# ========================================
# 可选依赖
# ========================================
# OpenClaw - 浏览器自动化（用于抓取Boss直聘真实岗位）
# 安装方法：npm install -g @openclaw/cli

